fix: left shift parenthesis edge case
When performing a left shift with a cast in the left hand side
expression, the rust parser requires parenthesis around the cast
operation.
This is also the case for the less-than operator for which this
edge-case was already caught.
This PR simply extends on that solution by adding an additional
left-shift check.